Refactor Server table
